federated_pneumonia_detection/requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Federated Learning
flwr[simulation]>=1.5.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Streamlit (for dashboard)
streamlit>=1.25.0

# Additional utilities
tqdm>=4.65.0
psutil>=5.9.0
tensorboard>=2.13.0
langchain-postgres>=0.0.1
psycopg[binary]>=3.1.0
